fix: unblock CI on issue-1088 - lint errors, notification-preferences collision, and build breaks
CI was red for several reasons, none related to the SMTP feature itself:

Lint (12 pre-existing errors from earlier merges, react-hooks/set-state-in-effect
and react/jsx-no-duplicate-props):
- add-issue-triage-board-ui.tsx, create-saved-filter-presets-page.tsx,
  notification-center/page.tsx, onboarding-wizard/page.tsx: suppressed
  set-state-in-effect with the established repo convention (see NavBar.tsx)
  for the intentional "hydrate from localStorage once after mount" pattern -
  keeps the SSR-matching first render, doesn't change behavior.
- implement-run-history-table-component.tsx: three <th> had duplicate
  className props (a merge artifact); kept the sort-aware dynamic one and
  discovered it referenced an undefined `sortField` - fixed to `sort.field`
  (the actual state, confirmed via the file's own useState<SortState>).
- notification-preferences-utils.test.ts, onboarding-wizard/onboarding-utils.test.ts:
  removed genuinely-unused imports.
- query-builder/page.tsx: updateGroupRecursive was referenced by 5 useCallback
  hooks before its own declaration; moved it above them as a useCallback with
  a named function expression (self-recursion via a const before its own
  initializer finishes isn't safe under this rule) and added it to their
  dep arrays.

Build - apps/web/src/app/notification-preferences-utils.ts was claimed by two
incompatible features from an earlier merge collision (flagged but not fixed
in the #1087/#1088 PR descriptions): add-notification-center-ui.tsx needs a
NotificationPreferences settings object + filterByPreferences/isInQuietHours,
while notification-center/page.tsx needs the per-channel NotificationPreference[]
array. Only the array version survived the merge. Reconstructed the missing
settings-object API from its three consumers (notification-preferences.tsx,
add-notification-center-ui.tsx) and the already-existing but quarantined spec
test (notification-center-integration.test.ts, which enumerates every function's
exact expected behavior) - all its assertions now pass, so it's been removed
from QUARANTINED_TESTS. Renamed the per-channel version's DEFAULT_PREFERENCES/
loadPreferences/savePreferences to DEFAULT_CHANNEL_PREFERENCES/
loadChannelPreferences/saveChannelPreferences (its only consumer,
notification-center/page.tsx, updated) to free those names for the restored
settings API, which two of the three real consumers actually need.

Build - a second, unrelated merge collision: campaign-milestone-timeline/page.tsx
imported two sibling files with a relative path one directory too shallow
(./add-campaign-milestone-timeline-visualizer instead of
../add-campaign-milestone-timeline-visualizer). Fixed both imports.

Build - api/auth/github/login/route.ts used the NextRequest type without
importing it. Added the import.

Build - api/runs/[id]/route.ts's fetch() call had two options objects (another
merge artifact) - JS only ever passed the second to native fetch anyway, but
TypeScript's fetch() overloads don't accept 3 arguments, so it never compiled.
Merged both into one options object (cache: 'no-store' + the 10s abort signal
are both clearly intended, not a pick-one situation).

Build - fixed a real bug this PR introduced: smtp-email.ts mixed nodemailer
(Node-only: net/tls/os/crypto) with the pure SmtpConfig/EmailMessage validation
that integrate-smtp-email-integration.tsx (a client component) imports from it,
so Next tried to bundle nodemailer for the browser and failed to resolve `tls`.
Split the pure types/validation/template builders into a new smtp-validation.ts;
smtp-email.ts now re-exports it and adds only the nodemailer-dependent
send/verify functions. The client component and smtp-adapter.ts now import
from smtp-validation.ts directly instead of smtp-email.ts.

Verification (all from apps/web, matching ci.yml's web job exactly):
- pnpm run lint: 0 errors, 1 pre-existing unrelated warning
- pnpm run test: full chain passes, including test:smtp-email and the
  newly-restored notification-center-integration.test.ts
- pnpm run build: exits 0, /integrations/smtp and all four new API routes
  compile and are listed in the route manifest
