+++ apps/web/src/app/campaign-milestone-timeline/page.tsx
@@ -1,7 +1,7 @@
 "use client";
 
-import CampaignMilestoneTimelineVisualizer from "./add-campaign-milestone-timeline-visualizer";
-import { mockMilestoneEvents } from "./campaign-milestone-mock-data";
+import CampaignMilestoneTimelineVisualizer from "../add-campaign-milestone-timeline-visualizer";
+import { mockMilestoneEvents } from "../campaign-milestone-mock-data";
 
 export default function CampaignMilestoneTimelinePage() {
   return (
