+++ apps/web/src/app/create-saved-filter-presets-page.tsx
@@ -25,6 +25,9 @@ export default function CreateSavedFilterPresetsPage() {
   const fileInputRef = useRef<HTMLInputElement>(null);
 
   useEffect(() => {
+    // Deferred to after mount so the initial render matches the server
+    // (empty) output before hydrating from localStorage.
+    // eslint-disable-next-line react-hooks/set-state-in-effect
     setPresets(readPresets());
     setHydrated(true);
   }, []);
