+++ apps/web/src/app/implement-run-history-table-component.tsx
@@ -224,11 +224,10 @@ export default function EnhancedRunHistoryTable({
               {visibleColumns.includes("id") && (
                 <th
                   className={`px-6 py-5 text-[10px] font-bold uppercase tracking-widest cursor-pointer transition-colors ${
-                    sortField === "id"
+                    sort.field === "id"
                       ? "text-blue-600 dark:text-blue-400"
                       : "text-zinc-400 hover:text-zinc-900 dark:hover:text-zinc-100"
                   }`}
-                  className="px-6 py-5 text-[10px] font-bold text-zinc-400 uppercase tracking-widest cursor-pointer hover:text-zinc-900 dark:hover:text-zinc-100 transition-colors"
                   aria-sort={getSortIndicator("id", sort).ariaSort}
                   onClick={() => toggleSort("id")}
                 >
@@ -248,11 +247,10 @@ export default function EnhancedRunHistoryTable({
               {visibleColumns.includes("duration") && (
                 <th
                   className={`px-6 py-5 text-[10px] font-bold uppercase tracking-widest text-right cursor-pointer transition-colors ${
-                    sortField === "duration"
+                    sort.field === "duration"
                       ? "text-blue-600 dark:text-blue-400"
                       : "text-zinc-400 hover:text-zinc-900 dark:hover:text-zinc-100"
                   }`}
-                  className="px-6 py-5 text-[10px] font-bold text-zinc-400 uppercase tracking-widest text-right cursor-pointer hover:text-zinc-900 dark:hover:text-zinc-100 transition-colors"
                   aria-sort={getSortIndicator("duration", sort).ariaSort}
                   onClick={() => toggleSort("duration")}
                 >
@@ -262,11 +260,10 @@ export default function EnhancedRunHistoryTable({
               {visibleColumns.includes("seedCount") && (
                 <th
                   className={`px-6 py-5 text-[10px] font-bold uppercase tracking-widest text-right cursor-pointer transition-colors ${
-                    sortField === "seedCount"
+                    sort.field === "seedCount"
                       ? "text-blue-600 dark:text-blue-400"
                       : "text-zinc-400 hover:text-zinc-900 dark:hover:text-zinc-100"
                   }`}
-                  className="px-6 py-5 text-[10px] font-bold text-zinc-400 uppercase tracking-widest text-right cursor-pointer hover:text-zinc-900 dark:hover:text-zinc-100 transition-colors"
                   aria-sort={getSortIndicator("seedCount", sort).ariaSort}
                   onClick={() => toggleSort("seedCount")}
                 >
