+++ apps/web/src/app/onboarding-wizard/onboarding-utils.test.ts
@@ -1,6 +1,5 @@
 import {
   ONBOARDING_STEPS,
-  completeStep,
   getProgressPercentage,
 } from './onboarding-utils';
